Set up formatter and run

+++ contracts/periphery/examples/ExampleComputeLiquidityValue.sol
@@ -1,6 +1,6 @@
 pragma solidity =0.6.6;
 
-import '../libraries/UniswapV2LiquidityMathLibrary.sol';
+import "../libraries/UniswapV2LiquidityMathLibrary.sol";
 
 contract ExampleComputeLiquidityValue {
     using SafeMath for uint256;
@@ -18,13 +18,14 @@ contract ExampleComputeLiquidityValue {
         uint256 truePriceTokenA,
         uint256 truePriceTokenB
     ) external view returns (uint256 reserveA, uint256 reserveB) {
-        return UniswapV2LiquidityMathLibrary.getReservesAfterArbitrage(
-            factory,
-            tokenA,
-            tokenB,
-            truePriceTokenA,
-            truePriceTokenB
-        );
+        return
+            UniswapV2LiquidityMathLibrary.getReservesAfterArbitrage(
+                factory,
+                tokenA,
+                tokenB,
+                truePriceTokenA,
+                truePriceTokenB
+            );
     }
 
     // see UniswapV2LiquidityMathLibrary#getLiquidityValue
@@ -32,16 +33,14 @@ contract ExampleComputeLiquidityValue {
         address tokenA,
         address tokenB,
         uint256 liquidityAmount
-    ) external view returns (
-        uint256 tokenAAmount,
-        uint256 tokenBAmount
-    ) {
-        return UniswapV2LiquidityMathLibrary.getLiquidityValue(
-            factory,
-            tokenA,
-            tokenB,
-            liquidityAmount
-        );
+    ) external view returns (uint256 tokenAAmount, uint256 tokenBAmount) {
+        return
+            UniswapV2LiquidityMathLibrary.getLiquidityValue(
+                factory,
+                tokenA,
+                tokenB,
+                liquidityAmount
+            );
     }
 
     // see UniswapV2LiquidityMathLibrary#getLiquidityValueAfterArbitrageToPrice
@@ -51,18 +50,17 @@ contract ExampleComputeLiquidityValue {
         uint256 truePriceTokenA,
         uint256 truePriceTokenB,
         uint256 liquidityAmount
-    ) external view returns (
-        uint256 tokenAAmount,
-        uint256 tokenBAmount
-    ) {
-        return UniswapV2LiquidityMathLibrary.getLiquidityValueAfterArbitrageToPrice(
-            factory,
-            tokenA,
-            tokenB,
-            truePriceTokenA,
-            truePriceTokenB,
-            liquidityAmount
-        );
+    ) external view returns (uint256 tokenAAmount, uint256 tokenBAmount) {
+        return
+            UniswapV2LiquidityMathLibrary
+                .getLiquidityValueAfterArbitrageToPrice(
+                    factory,
+                    tokenA,
+                    tokenB,
+                    truePriceTokenA,
+                    truePriceTokenB,
+                    liquidityAmount
+                );
     }
 
     // test function to measure the gas cost of the above function
@@ -72,10 +70,8 @@ contract ExampleComputeLiquidityValue {
         uint256 truePriceTokenA,
         uint256 truePriceTokenB,
         uint256 liquidityAmount
-    ) external view returns (
-        uint256
-    ) {
-        uint gasBefore = gasleft();
+    ) external view returns (uint256) {
+        uint256 gasBefore = gasleft();
         UniswapV2LiquidityMathLibrary.getLiquidityValueAfterArbitrageToPrice(
             factory,
             tokenA,
@@ -84,7 +80,7 @@ contract ExampleComputeLiquidityValue {
             truePriceTokenB,
             liquidityAmount
         );
-        uint gasAfter = gasleft();
+        uint256 gasAfter = gasleft();
         return gasBefore - gasAfter;
     }
 }
